requirements.txt
fastapi==0.88.0
pydantic==1.10.4
SQLAlchemy==1.4.45
uvicorn==0.20.0
Werkzeug==2.2.2
python_multipart==0.0.6
PyJWT==1.7.0